+++ restaurant.client/src/Cart/CartItemsList.jsx
@@ -1,20 +1,9 @@
 import React from "react";
 import { Row } from "react-bootstrap";
 import CartItem from "./CartItem";
-import { deleteMealFromOrder } from "../ApiCall";
 
 const CartItemsList = (props) => {
   const meals = props.meals;
-
-  const handleMealDelete = async (rowId) => {
-    try {
-      const result = await deleteMealFromOrder(rowId);
-      const updatedMeals = meals.filter(item => item.rowId !== rowId);
-      props.setMeals(updatedMeals);
-    } catch (error) {
-      console.error("Error in handleDelete: ", error);
-    }
-  };
 
   return (
     <Row>
@@ -28,7 +17,7 @@ const CartItemsList = (props) => {
             weight={item.weight}
             amount={item.amount}
             handleTotalUptade={props.handleTotalUptade}
-            handleMealDelete={handleMealDelete}
+            handleMealDelete={props.handleMealDelete} 
           />
         </div>
       ))}
